Refactor for testing projects
