--- src/copyright.ts
@@ -1,3 +1,0 @@
-export const replaceWithCurrentYear = (input: string, match: string): string => {
-    return input.replace(match, new Date().getFullYear().toString());
-};
